+++ Moco/Moco/RegisterTypes_osimMoco.cpp
@@ -33,16 +33,18 @@
 #include "MocoBounds.h"
 #include "MocoCasADiSolver/MocoCasADiSolver.h"
 #include "MocoControlBoundConstraint.h"
-#include "MocoCost/MocoControlCost.h"
-#include "MocoCost/MocoControlTrackingCost.h"
-#include "MocoCost/MocoCost.h"
-#include "MocoCost/MocoJointReactionCost.h"
-#include "MocoCost/MocoMarkerFinalCost.h"
-#include "MocoCost/MocoMarkerTrackingCost.h"
-#include "MocoCost/MocoOrientationTrackingCost.h"
-#include "MocoCost/MocoStateTrackingCost.h"
-#include "MocoCost/MocoSumSquaredStateCost.h"
-#include "MocoCost/MocoTranslationTrackingCost.h"
+#include "MocoGoal/MocoControlGoal.h"
+#include "MocoGoal/MocoControlTrackingGoal.h"
+#include "MocoGoal/MocoGoal.h"
+#include "MocoGoal/MocoInitialActivationGoal.h"
+#include "MocoGoal/MocoJointReactionGoal.h"
+#include "MocoGoal/MocoMarkerFinalGoal.h"
+#include "MocoGoal/MocoMarkerTrackingGoal.h"
+#include "MocoGoal/MocoOrientationTrackingGoal.h"
+#include "MocoGoal/MocoPeriodicityGoal.h"
+#include "MocoGoal/MocoStateTrackingGoal.h"
+#include "MocoGoal/MocoSumSquaredStateGoal.h"
+#include "MocoGoal/MocoTranslationTrackingGoal.h"
 #include "MocoInverse.h"
 #include "MocoParameter.h"
 #include "MocoProblem.h"
@@ -62,18 +64,21 @@ static osimMocoInstantiator instantiator;
 
 OSIMMOCO_API void RegisterTypes_osimMoco() {
     try {
-        Object::registerType(MocoFinalTimeCost());
+        Object::registerType(MocoFinalTimeGoal());
         Object::registerType(MocoWeight());
         Object::registerType(MocoWeightSet());
-        Object::registerType(MocoStateTrackingCost());
-        Object::registerType(MocoMarkerTrackingCost());
-        Object::registerType(MocoMarkerFinalCost());
-        Object::registerType(MocoControlCost());
-        Object::registerType(MocoSumSquaredStateCost());
-        Object::registerType(MocoControlTrackingCost());
-        Object::registerType(MocoJointReactionCost());
-        Object::registerType(MocoOrientationTrackingCost());
-        Object::registerType(MocoTranslationTrackingCost());
+        Object::registerType(MocoStateTrackingGoal());
+        Object::registerType(MocoMarkerTrackingGoal());
+        Object::registerType(MocoMarkerFinalGoal());
+        Object::registerType(MocoControlGoal());
+        Object::registerType(MocoSumSquaredStateGoal());
+        Object::registerType(MocoControlTrackingGoal());
+        Object::registerType(MocoInitialActivationGoal());
+        Object::registerType(MocoJointReactionGoal());
+        Object::registerType(MocoOrientationTrackingGoal());
+        Object::registerType(MocoTranslationTrackingGoal());
+        Object::registerType(MocoPeriodicityGoalPair());
+        Object::registerType(MocoPeriodicityGoal());
         Object::registerType(MocoBounds());
         Object::registerType(MocoInitialBounds());
         Object::registerType(MocoFinalBounds());